requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
cryptography==42.0.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
scikit-learn==1.4.0
numpy==1.26.3
joblib==1.3.2
tenacity==8.2.3
httpx==0.26.0
python-dotenv==1.0.0